third_party/blink/web_tests/external/wpt/html/syntax/parsing/inhead-noscript-head-expected.txt

This is a testharness.js-based test.
[FAIL] When the scripting flag is disabled, a head start tag in "in head noscript" mode should be ignored
  assert_equals: expected "[object HTMLStyleElement]" but got "null"
Harness: the test ran to completion.
